renamed package, update gradle.build
